Update .gitignore to exclude .pnpm-store and TypeScript build info files; refactor PredictionEntryModal to improve type safety and remove unused code; update contest and lineup routes to enforce type definitions for better maintainability; bump hono package version to 4.10.3 in server configuration.

+++ client/src/hooks/useContestPredictionData.ts
@@ -116,7 +116,7 @@ export function useContestPredictionData(options: UseContestPredictionDataOption
       supply &&
       supply > 0n &&
       totalSpectatorCollateral &&
-      totalSpectatorCollateral > 0n
+      (totalSpectatorCollateral as bigint) > 0n
     ) {
       impliedWinnings = (balance * (totalSpectatorCollateral as bigint)) / supply;
       impliedWinningsFormatted = formatUnits(impliedWinnings, 18);
